chore(VLE): Merge themeController code into vleController
#527
