+++ src/assets/wise5/themes/default/theme.js
@@ -1,11 +1,9 @@
 'use strict';
 
-import ThemeController from './themeController';
 import './themeComponents';
 
 import './js/webfonts';
 
-const themeModule = angular.module('theme', ['theme.components'])
-    .controller('ThemeController', ThemeController);
+const themeModule = angular.module('theme', ['theme.components']);
 
 export default themeModule;
